requirements.txt
fastapi==0.88.0
uvicorn==0.20.0

pydantic~=1.10.4
neo4j~=5.3.0
python-dotenv~=0.21.0